Add support for comments
This moves the printing and consuming of tokens to one method, which
also checks if we're ahead of the token stream and have to catch up
comments.
Also adds a test for commented "Hello, World!"

Known issues:
Inline "multi-line" comments (`print("like " + /* for example */ +
"this")`) don't work, since "newline after a multi-line comment" is
currently hard-coded instead of being read from the input.