Issue #7: Add LineBreakStrategy + dumb strategy
The LineBreakStrategy decides where to place the next line break. The
dumb strategy places the line break as late as possible; I plan to add a
more sophisticated strategy later which will try to ensure that the next
line’s indentation level isn’t too big.

At the moment, the dumb strategy is still a bit *too* dumb and inserts
line breaks even when the tokens forbid a line break (by giving an
indentBefore/postIndent of null, e. g. before a ‘,’ or ‘(’).

Also fixes #9.

Also, fixes the AntlrToken alias (was AntrlToken – check the ‘lr’ if you
don’t see the difference) – that was quite confusing when it didn’t work
when I typed it manually.