FormattingWriter: fix bug closing contexts too early
In the following situation, <> opens and closes an empty context, and ()
opens and closes an actual context associated with these tokens:

    <void method(
        "Multiline because of doc string"
        Anything param)> {

This would be formatted as

    void method(
        "Multiline because of doc string"
    Anything param) {

because closeContext(>) closed the context immediately, cleaning (’s
opening context from the tokenStack even though the tokens Anything,
param and ) should still be written with that context.

This bug has been around for a while, but surfaced through previous
commit a6982bd. Fortunately, the fix is easy since I already had to deal
with this case for closing tokens (like the )).