Implement sequenced arguments / sequence enumerations
    value v = { a, b, c };

This also removes the indentAfter = Indent(1) of an equals sign, because
that led to a double indent of a sequence enumeration spread across
multiple lines (one from the equals sign, one from the opening brace).
It seems to make sense, but I believe that in all cases where a line
break can occur, there is some other construct that will provide the
indent (for example, MemberOp’s indentBefore = Indent(1), or the indents
of argument lists’ opening parentheses).