Implement for clauses
Consisting of a ForClause and a ValueIterator. This also broke some
assumptions I had made earlier – namely, that a specifier expression’s
mainToken is always a ‘=’ (can also be ‘in’), and that statements always
end with a ‘;’ (for clauses are also statements). Also, sometimes the
AST contains ValueModifiers that aren’t in the source file (null token),
and TypedDeclarations with a null annotationList (instead of an empty
one; bug? I don’t really care, it’s easy to fix).