Implement Exists, Nonempty
With this, FormattingVisitor.ceylon can successfully be formatted.
Running the FormattingVisitor with and over the formatted code
introduces no further changes, which means that the formatter produced
valid code. There are still some ugly things in the output – and of
course, the FormattingVisitor doesn’t use all language features, so the
formatter isn’t complete yet – but this is still a big step.