Fix #30 for imports of keywords
And if you want to import a *class* called “in” as \Iin, fuck you. Known
corner case, WONTFIX.

In more pleasant news, this brings us very *very* close to #29,
dogfooding. The formatted code compiles and runs fine, and the only
difference between the once and twice formatted code is a missing blank
line.