Don’t generate the LineBreaks class
The LineBreaks class now refines string to mean the line breaking
sequence, and the instances set it correctly. This allows us to remove
the line breaking sequence determination from FormattingWriter, which is
a lot cleaner.

Also adds the feature of non-generated enums to source-gen. LineBreaks
is now a hand-written file that isn’t generated, but source-gen will
still use the knowledge of the cases when parsing options (instead of
delegating to a parseLineBreaks function).