Fix really stupid bug in DefaultLineBreaks
FormattingWriter.LineBreak refers to the queue element type.
LineBreak used to refer to that, too, but the behavior of the import on
top changed (ceylon/ceylon-spec#989), and now it refers to the generated
LineBreak (from options, lf/crlf/os).

This took decidedly too long to find, it’s a really stupid mistake :-(