Use String.lines in FormattingWriter
I investigated a bit with VisualVM, and it seems that splitting a string
into lines is one of the hot spots. At the moment, String.lines isn’t
optimized, but even if it’s never hand-optimized, the JVM might still
find it marginally easier to optimize several uses of String.lines than
several occurences of String.split('\n'.equals).