Optimize for one-line tokens in FormattingWriter.write()
Investigation with VisualVM revealed that splitting tokens into several
lines in FormattingWriter.write() (for indenting them) is quite
expensive, and one of the hot spots of the formatter. Since the vast
majority of tokens have only one line, we can avoid that most of the
time.

In a test where I formatted FormattingWriter.ceylon 1000 times (doing
the parsing outside the loop), this increased performance from
25203664443ns to 23438177240ns (25s → 23s).