Fix bug with line breaks after a line comment
If there is, for example, an opening brace ('{') after a line comment,
the intersection of the allowed line breaks after the line comment
(1...) and the allowed line breaks before the other token (0..0) is
empty, which would crash the formatter.

Found in ceylon/ceylon.ast@ee6b7e2
(source/ceylon/ast/redhat/Invocation.ceylon).