Implement profiles (closes #74)
A profile is a configuration file with the name “profile.$name”; options
are loaded from the section ‘formatter’. The profile can be set using
the ‘--profile=$name’ command line option or the ‘formattool.profile’
Ceylon config option (in the regular config). (Specifying a file path in
the ‘--profile‘ command line option is not (yet?) supported.)

CC ceylon/ceylon-ide-eclipse#987.