Fix ceylon/ceylon-ide-eclipse#1157
I don’t know why named argument lists had an indentBefore of 1, it looks
stupid.

More generally, though, the IDE should probably remove leading
indentation from the first formatted “unit”, as there are of course also
legitimate uses of a nonzero indentBefore (although this is only
critical if the IDE’s FindNodeVisitor also yields these nodes – for
example, an `extends` clause has an indentBefore of 2 (by default), but
if you select it, it seems that it’s always formatting the entire
class).