Fix formatting typechecked anonymous arguments
Anonymous named arguments like in

    print {
        "Hello, World!";
    }

are parsed as specified arguments with a ‘null’ identifier and a
specifier expression with ‘null’ token. However, it seems that the
typechecker later adds a synthetic identifier for the corresponding
parameter. This isn’t relevant for the command line formatter (since we
only use the parser), but it is relevant for the IDE, which formats a
completely typechecked AST. So we have to detect those synthetic
identifiers (with ‘null’ tokens).