Simplify spaceAfterParamListClosingParen option
Exposing the internal “weight” system like that was not a good idea…
it’s far too complicated. Also, even if you understand it, it’s probably
not very useful unless you know all the other weights too, which are not
exposed.

See some discussion (with examples) on the ceylon-dev mailing list:
https://groups.google.com/d/msg/ceylon-dev/i-AF_HLnR04/4RtXXsgtZa8J
Thanks @quintesse for input!

Also updates the tests to include all five tokens that (to my knowledge)
can occur after a parameter list’s closing parenthesis.

Remember to update your old profiles (the default value was -10, which
now corresponds to true).