Update for ceylon/ceylon-spec@ef32121
The formatter now handles the syntax

    case (x1 | x2)

correctly. The old syntax

    case (x1, x2)

is not supported – we can’t support it, since the separating token isn’t
recorded. This means that when the old syntax is encountered, the
formatter crashes (unexpected token). Sorry everyone.

(The formatter source code itself doesn’t have any multiple-value switch
cases, so this has no effect on dogfooding.)