Use switch clause parentheses from AST
As of ceylon/ceylon-spec@57304d7, the grammar no longer throws away the
parentheses around a switch clause’s expression, instead recording them
in the switch clause’s end token. The lparen is overwritten by the
rparen, but we can use the rparen instead of constructing it manually.

(Note: This breaks the ceylon.ast tests if ceylon/ceylon.ast@fab1bcf
isn’t pulled in as well.)