Don’t overwrite AST models
The parameter lists of anonymous functions are formatted slightly
differently from those of regular functions because, since they don’t
have an identifier, they look stupid with indentation before the opening
parenthesis. In order to be able to make this distinction within
`visitParameterList`, we used to set the model of a parameter list to a
special marker object in some cases (though I forgot to do it when later
adding support for constructors, #98). That was a really stupid idea,
already in general, but especially for a tool intended for IDE
integration. The much better solution is to call a special visit
function in the special case (within `visitFunctionArgument`). The
author of the original code has been firmly admonished.

Fixes #107.