Fix one part of #115
DefaultLineBreaks would, upon encountering a multi-line token, always
break at that token, regardless of any line breaks before it. Instead,
we should stop walking the queue, but not necessarily return early;
instead, only do what we would otherwise do when the queue is done (this
*does* include an early return if we weren’t even over the max line
length yet), and then proceed as normal.

Note that this alone doesn’t actually improve the situation: It
exchanges doc strings being pulled up to the preceding token with doc
strings pulling up the succeeding token to them. That part is fixed in
the following commit.