Fix another part of #115
With the previous commit but without this one, lines *after* doc strings
would be “pulled up” (instead of the doc strings being pulled up to the
lines before them). This is because writeLine() conflated two similar,
but not identical, conditions:

1. When there is a multi-line token, alignment of that token needs to be
   adjusted. However, this adjustment needs to happen relative to the
   indentation of the line on which the literal starts, without taking
   into account any additional indentation that this line introduced.
   For this reason, if the last token for this line is a multi-line
   token (a multi-line token can never be an earlier token for its line,
   since it naturally terminates the line – thus, looking just at the
   last token is an optimization), indentation handling is deferred
   until after the multi-line token is done.

2. If the very last queue element is a multi-line literal, then
   logically the line didn’t end with that token. Consider for example:

       print("Multi-line
              string!");

   The line continues after the closing quote of the string, and spacing
   between the closing quote and the closing parenthesis matters. For
   this reason, we should not treat this like the end of a line.

Previously, the same condition was used for these two cases, but the
condition is actually slightly different: 1 is based on the last token
of the line (ignoring other elements), 2 is based on the last element.
Those are different, for instance, when there’s an explicit line break
after the last token – previously, this was ignored, which caused the
“pulling up” of whatever came after a multi-line token.