Fix missing indentation of type param/arg lists
The formatter produced

    foo<Bar,
    Baz>();

instead of

    foo<Bar,
        Baz>();

This is because of an (as far as I know, undocumented) rule where
writeToken() would only respect indentBefore and indentAfter if a line
break is allowed before/after the token. This sounded reasonable back
when I wrote it, and it still sounds kinda reasonable now, but as it
turns out, there’s half a dozen places in FormattingWriter where I give
an indentAfter or indentBefore even though a line break isn’t allowed
after/before that token (one of them is
writeTypeArgumentOrParameterList), so clearly there are places where
it’s reasonable to introduce indentation even though no line breaks are
allowed *immediately* after the token.

One test needed to be updated for this, as it actually contained the
missing indentation in a line-broken type argument list.