Make direction and stacking of indentation orthogonal
Previously, indentation after a token always stacked, and indentation
before a token never did. This model worked fairly well, but its
limitations were clear in some cases, and to solve these cases, I tacked
on some fairly ugly hacks like “nextIndentBefore” and
“indentAfterOnlyWhenLineBreak”.

Instead, the direction and stacking of indentation should be orthogonal
notions: now, indentation in either direction may stack always, never,
or only if it’s actually applied (at the start or end of a line,
respectively). This is a much cleaner and more powerful model, which
will also allow a proper solution to #37 / #105.

Some of this amounts to a partial rewrite of some parts of
FormattingWriter (see #103); notably, Token is now an enumerated type.
This necessitates the full rewrite of the DefaultLineBreaks test; which
now uses only the FormattingWriter interface instead of creating tokens
directly. This hopefully means it can be left alone in the future when
the FormattingWriter changes further.