Fix #105: indentation of `=`/`=>`
Previously, the indentation of `=` and `=>` could be fairly ugly in some
quite common scenarios, as can be seen in the test for #104. The
solution, as discussed in #105, is that the indentation both before and
after an `=` or `=>` should stack if and only if it is actually applied.
Support for this was added to FormattingWriter in the previous commits.

This supersedes the solution of #37, and thus the tests for #37 are
removed completely. New tests for #105 will be added in subsequent
commits.

One remaining problem is that currently the desired indentation behavior
after an `=` or `=>` only occurs with the option
+indentationAfterSpecifierExpressionStart set to stack, which is not
currently the default. This needs some further investigation for optimal
interoperation with the Eclipse IDE’s Fix Indentation action.