Clean up build files
Certain ceylon.formatter build steps require libraries (ant-contrib,
OSGi libraries) that are included in the main Ceylon source tree, but
are not shipped with the distribution. This commit makes it clearer in
the properties file that these are separate, and that the main build can
be done by updating dist.dir in build.properties without needing any
ceylon.root.dir (which a shipped distribution does not provide).

(This also makes it easier to produce reproducible builds of
ceylon.formatter with the same script as for ceylon.ast, since in both
cases you can now simply update dist.dir to point to the downloaded
distribution.)