Work around ceylon/ceylon#6914
This slightly widens the type of testCases to remove the union type,
which is enough to fix the backend error. I could also declare the
iterables to be possibly-empty – I don’t need the nonemptiness for
anything – but I like the new type enough that I might even keep it
after the compiler bug is fixed: I think it’s exactly the right type for
testCases, as precise as possible while still making sense (it doesn’t
matter that the inner keys are only [String] and [String,String] right
now, the type they need to have to match parseTranslations’ result is
String[]).