Rewrite single-line-ness and switch-else-ness
We maintain a set of nodes that are allowed to be a single line, and
nodes that are part of a switch’s else. This information is propagated
through nodes in certain cases. visitBody() may skip the line-break
enforcement around comments for single-line blocks if the block is in
the first set, and visitElseClause() enforces a line break before the
else token if the clause is in the second set.

The set of nodes that are allowed to be a single line is only sparsely
populated; many kinds of blocks that could previously be a single line
(type and function bodies, try/catch/finally, …) are not eligible, and
will be forced into several lines if anyone actually used them in
single-line form. (It’s worth pointing out that the formatter never
produces the single-line form by itself, it merely allows it to remain
in place in certain cases.)

Fixes #130.