Add support for constants in module descriptors
Constants can now be declared in module descriptors (they look like
ordinary values) and be used as the version (instead of a string
literal). This commit adds support for that.

Since constants and module descriptors can be interleaved, but the
parser stores them into two separate lists, we have to restore their
order based on the token index. The code to do this is mostly copied
from casesList, where we already have to do the same thing. (I opened an
issue for that back in the day, but I don’t think that’s necessary to do
this time, the workaround works well enough.)

Fixes #140.