Add support for new destructuring syntax
Previously, a Destructure node consisted of an optional ValueModifier, a
Pattern and a SpecifierExpression, so it was supported by the default
visitor behavior without an explicit visitDestructure method. With
ceylon/ceylon@db53d11903, the `value` token is no longer part of the
Destructure, but of a new LetStatement node (and just recorded as token,
not as ValueModifier), so this is no longer enough: we need to implement
visitLetStatement to write the main token (`value` or `let`), and, in
case of `let`, the parentheses and the commas.

Fixes #142.