Support anonymous functions without parameter lists
Syntactically, the new anonymous function syntax is simply an anonymous
function with an empty list of parameter lists (anonymous functions
could already have multiple parameter lists), so all we need to do is to
not require the space in front of the fat arrow if there are no
parameter lists, since in that case the space isn’t necessary.

Part of #147.