Add support for pipe (|>) and fish (>|>) expressions
Instead of treating them like ordinary binary expressions (which would
require no specific support from ceylon.formatter), the parser desugars
these expressions into InvocationExpressions. To format them correctly,
we have to detect them based on their token type and then extract just
the right bits out of the InvocationExpression which correspond to the
original code.

The snippet to format a binary operator with correct indentation is
extracted from visitBinaryOperatorExpression() into
writeBinaryOperator() in Patterns.ceylon so that it can be used in
visitPipeExpression() and visitFishExpression() as well.

Fixes #151, fixes #152. (The operators are similar enough that we might
as well do this in one commit, I think.)