Fix implementation of pipe and fish operators
In commit 494c42bad1, I published code which simply does not compile:
the FormattingWriter argument to writeBinaryOperator() is missing. I’m
not sure how this happened, but I suspect that I first wrote the
implementation with copy+pasting the writeBinaryOperator() snippet (and
also tested it like this), and then only extracted the function at the
end, and didn’t run the tests again or even tried to compile the module.

I also didn’t use the extracted function in
visitBinaryOperatorExpression(), which is where it was extracted from.
Fix both those issues.

Pointed out by @CPColin – thanks!